// src/types/types.ts
export interface CardData {
    name: string;
    link: string;
}

export interface UserInfoData {
    name: string;
    about: string;
}

export interface FormConfig {
    formSelector: string;
    inputSelector: string;
    submitButtonSelector: string;
    inactiveButtonClass: string;
    inputErrorClass: string;
    errorClass: string;
}